[RMATE-01] - organize the first structure expectation
