+++ app/src/main/java/com/rodcollab/readermate/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.rodcollab.readermate.app
+package com.rodcollab.readermate
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.rodcollab.readermate.app.theme.ReaderMateTheme
+import com.rodcollab.readermate.ui.theme.ReaderMateTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
