+++ app/src/main/java/com/rodcollab/readermate/books/BookDto.kt
@@ -1,4 +1,4 @@
-package com.rodcollab.readermate.core.network.model
+package com.rodcollab.readermate.books
 
 /**
  * Model related to the google books api
