+++ app/src/main/java/com/rodcollab/readermate/books/BookRecord.kt
@@ -1,4 +1,4 @@
-package com.rodcollab.readermate.core.model
+package com.rodcollab.readermate.books
 
 /**
  * Books that the user will choose to read
