+++ app/src/main/java/com/rodcollab/readermate/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.rodcollab.readermate.app.theme
+package com.rodcollab.readermate.ui.theme
 
 import androidx.compose.ui.graphics.Color
